Fixed some cursor warping issues Improved window handler and functions to contain it's parent screen size for better multi-screen support Improved set_region/subregion_scaled performance when downscaling changed mlx's fps cap to help with FPS desyncs and mitigate screen tearing "Improved" (removed op_to functions and added vec_mod and missing operation variants) some VecLib stuff

+++ lib/VecLibC/include/modules/types/veclc_s_vec2xn.h
@@ -52,7 +52,7 @@ t_vec2x2	vec2x2(t_vec2 p1, t_vec2 p2);
 /**
  * Creates a group of two 2D vectors containing doubles.
  */
-t_vec2x2	vec2x2_xy(int x1, int y1, int x2, int y2);
+t_vec2x2	vec2x2_xy(double x1, double y1, double x2, double y2);
 
 /**
  * Creates a group of three 2D vectors containing doubles.
